Fix documentation building errors (#1360)

+++ examples/waveform/plot_fd_td.py
@@ -14,7 +14,7 @@ sptilde, sctilde = waveform. get_fd_waveform(approximant="TaylorF2",
                              mass1=6, mass2=6, delta_f=1.0/4, f_lower=40)
 
 # FFT it to the time-domain  
-tlen = 1.0 / hp.delta_t / sptilde.delta_f
+tlen = int(1.0 / hp.delta_t / sptilde.delta_f)
 sptilde.resize(tlen/2 + 1)
 sp = types.TimeSeries(types.zeros(tlen), delta_t=hp.delta_t)                            
 fft.ifft(sptilde, sp)
